Refactor out day month year fields into immutable function components.
